+++ Modules/Microsoft365DSC/DSCResources/MSFT_AADUser/MSFT_AADUser.psm1
@@ -513,6 +513,7 @@ function Set-TargetResource
 
             $CreationParams.Add('UserId', $UserPrincipalName)
             Update-MgUser @CreationParams
+            $userId = (Get-MgUser -UserId $UserPrincipalName).Id
         }
         else
         {
@@ -559,6 +560,7 @@ function Set-TargetResource
             $CreationParams.Add('MailNickName', $UserPrincipalName.Split('@')[0])
             Write-Verbose -Message "Creating new user with values: $(Convert-M365DscHashtableToString -Hashtable $CreationParams)"
             $user = New-MgUser @CreationParams
+            $userId = $user.Id
         }
 
         #region Assign Licenses
@@ -611,7 +613,7 @@ function Set-TargetResource
 
                         throw "Cannot add user $UserPrincipalName to group '$memberOfGroup' because it is a dynamic group"
                     }
-                    New-MgGroupMember -GroupId $group.Id -DirectoryObjectId (Get-MgUser -UserId $UserPrincipalName).Id
+                    New-MgGroupMember -GroupId $group.Id -DirectoryObjectId $userId
                 }
             }
             else
@@ -642,14 +644,14 @@ function Set-TargetResource
 
                             throw "Cannot add user $UserPrincipalName to group '$($_.InputObject)' because it is a dynamic group"
                         }
-                        New-MgGroupMember -GroupId $group.Id -DirectoryObjectId (Get-MgUser -UserId $UserPrincipalName).Id
+                        New-MgGroupMember -GroupId $group.Id -DirectoryObjectId $userId
                     }
                     else
                     {
 
                         # Group that user is a member of is not present in MemberOf, remove user from group
                         # (no need to test for dynamic groups as they are ignored in Get-TargetResource)
-                        Remove-MgGroupMemberDirectoryObjectByRef -GroupId $group.Id -DirectoryObjectId (Get-MgUser -UserId $UserPrincipalName).Id
+                        Remove-MgGroupMemberDirectoryObjectByRef -GroupId $group.Id -DirectoryObjectId $userId
                     }
                 }
             }
@@ -676,7 +678,6 @@ function Set-TargetResource
             foreach ($roleDifference in $diffRoles)
             {
                 $roleDefinitionId = (Get-MgBetaRoleManagementDirectoryRoleDefinition -Filter "DisplayName eq '$($roleDifference.InputObject)'").Id
-                $userId = (Get-MgUser -UserId $UserPrincipalName).Id
 
                 # Roles to remove
                 if ($roleDifference.SideIndicator -eq '=>')